pyproject.toml
[project]
name = "pdf-to-es"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pymupdf>=1.26.3",
    "openai>=1.35.0",
    "anthropic>=0.30.0",
    "python-dotenv>=1.0.1",
    "pymupdf4llm>=0.0.27",
    "markdown>=3.8.2",
    "weasyprint>=66.0",
]

[project.scripts]
pdf-to-es = "main:main"
